requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
dj-database-url==2.1.0
Django==5.0.4
djangorestframework==3.15.1
environs==11.0.0
idna==3.7
marshmallow==3.21.1
packaging==24.0
pillow==10.3.0
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
sqlparse==0.5.0
transliterate==1.10.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
